+++ packages/expo-updates/ios/EXUpdates/EXUpdatesErrorRecovery.m
@@ -15,6 +15,7 @@ typedef NS_ENUM(NSInteger, EXUpdatesErrorRecoveryTask) {
   EXUpdatesErrorRecoveryTaskCrash
 };
 
+static NSString * const EXUpdatesErrorLogFile = @"expo-error.log";
 static NSInteger const EXUpdatesErrorRecoveryRemoteLoadTimeoutMs = 5000;
 
 @interface EXUpdatesErrorRecovery ()
@@ -26,6 +27,7 @@ static NSInteger const EXUpdatesErrorRecoveryRemoteLoadTimeoutMs = 5000;
 @property (nonatomic, assign) NSInteger remoteLoadTimeout;
 
 @property (nonatomic, strong) dispatch_queue_t errorRecoveryQueue;
+@property (nonatomic, strong, nullable) dispatch_queue_t diskWriteQueue;
 
 @property (nonatomic, strong) NSMutableArray *encounteredErrors;
 
@@ -39,10 +41,12 @@ static NSInteger const EXUpdatesErrorRecoveryRemoteLoadTimeoutMs = 5000;
 - (instancetype)init
 {
   return [self initWithErrorRecoveryQueue:dispatch_queue_create("expo.controller.errorRecoveryQueue", DISPATCH_QUEUE_SERIAL)
+                           diskWriteQueue:nil
                         remoteLoadTimeout:EXUpdatesErrorRecoveryRemoteLoadTimeoutMs];
 }
 
 - (instancetype)initWithErrorRecoveryQueue:(dispatch_queue_t)errorRecoveryQueue
+                            diskWriteQueue:(nullable dispatch_queue_t)diskWriteQueue
                          remoteLoadTimeout:(NSInteger)remoteLoadTimeout
 {
   if (self = [super init]) {
@@ -57,6 +61,7 @@ static NSInteger const EXUpdatesErrorRecoveryRemoteLoadTimeoutMs = 5000;
     _isWaitingForRemoteUpdate = NO;
     _rctContentHasAppeared = NO;
     _errorRecoveryQueue = errorRecoveryQueue;
+    _diskWriteQueue = diskWriteQueue;
     _remoteLoadTimeout = remoteLoadTimeout;
     _encounteredErrors = [NSMutableArray new];
   }
@@ -71,11 +76,13 @@ static NSInteger const EXUpdatesErrorRecoveryRemoteLoadTimeoutMs = 5000;
 - (void)handleError:(NSError *)error
 {
   [self _startPipelineWithEncounteredError:error];
+  [self writeErrorOrExceptionToLog:error];
 }
 
 - (void)handleException:(NSException *)exception
 {
   [self _startPipelineWithEncounteredError:exception];
+  [self writeErrorOrExceptionToLog:exception];
 }
 
 - (void)notifyNewRemoteLoadStatus:(EXUpdatesRemoteLoadStatus)newStatus
@@ -280,6 +287,85 @@ static NSInteger const EXUpdatesErrorRecoveryRemoteLoadTimeoutMs = 5000;
   RCTSetFatalExceptionHandler(_previousFatalExceptionHandler);
   _previousFatalErrorHandler = nil;
   _previousFatalExceptionHandler = nil;
+}
+
+# pragma mark - error persisting
+
++ (nullable NSString *)consumeErrorLog
+{
+  NSString *errorLogFilePath = [[self class] _errorLogFilePath];
+  NSData *data = [NSData dataWithContentsOfFile:errorLogFilePath options:kNilOptions error:nil];
+  if (data) {
+    NSError *err;
+    if (![NSFileManager.defaultManager removeItemAtPath:errorLogFilePath error:&err]) {
+      NSLog(@"Could not delete error log: %@", err.localizedDescription);
+    }
+    return [[NSString alloc] initWithData:data encoding:NSUTF8StringEncoding];
+  } else {
+    return nil;
+  }
+}
+
+- (void)writeErrorOrExceptionToLog:(id)errorOrException
+{
+  dispatch_async(_diskWriteQueue ?: dispatch_get_global_queue(DISPATCH_QUEUE_PRIORITY_DEFAULT, 0), ^{
+    NSString *serializedError;
+    if ([errorOrException isKindOfClass:[NSError class]]) {
+      serializedError = [NSString stringWithFormat:@"Fatal error: %@", [self _serializeError:(NSError *)errorOrException]];
+    } else if ([errorOrException isKindOfClass:[NSException class]]) {
+      serializedError = [NSString stringWithFormat:@"Fatal exception: %@", [self _serializeException:(NSException *)errorOrException]];
+    } else {
+      return;
+    }
+
+    NSData *data = [serializedError dataUsingEncoding:NSUTF8StringEncoding];
+    NSString *errorLogFilePath = [[self class] _errorLogFilePath];
+    if ([NSFileManager.defaultManager fileExistsAtPath:errorLogFilePath]) {
+      NSFileHandle *fileHandle = [NSFileHandle fileHandleForWritingAtPath:errorLogFilePath];
+      [fileHandle seekToEndOfFile];
+      [fileHandle writeData:data];
+      [fileHandle closeFile];
+    } else {
+      NSError *err;
+      if (![data writeToFile:[[self class] _errorLogFilePath] options:NSDataWritingAtomic error:&err]) {
+        NSLog(@"Could not write fatal error to log: %@", err.localizedDescription);
+      }
+    }
+  });
+}
+
+- (NSString *)_serializeException:(NSException *)exception
+{
+  return [NSString stringWithFormat:@"Time: %f\nName: %@\nReason: %@\n\n",
+    [NSDate date].timeIntervalSince1970 * 1000,
+    exception.name,
+    exception.reason];
+}
+
+- (NSString *)_serializeError:(NSError *)error
+{
+  NSString *localizedFailureReason = error.localizedFailureReason;
+  NSError *underlyingError = error.userInfo[NSUnderlyingErrorKey];
+
+  NSMutableString *serialization = [[NSString stringWithFormat:@"Time: %f\nDomain: %@\nCode: %li\nDescription: %@",
+                                     [[NSDate date] timeIntervalSince1970] * 1000,
+                                     error.domain,
+                                     (long)error.code,
+                                     error.localizedDescription] mutableCopy];
+  if (localizedFailureReason) {
+    [serialization appendFormat:@"\nFailure Reason: %@", localizedFailureReason];
+  }
+  if (underlyingError) {
+    [serialization appendFormat:@"\n\nUnderlying Error:\n%@", [self _serializeError:underlyingError]];
+  }
+  [serialization appendString:@"\n\n"];
+  return serialization;
+}
+
++ (NSString *)_errorLogFilePath
+{
+  NSURL *applicationDocumentsDirectory = [[NSFileManager.defaultManager URLsForDirectory:NSApplicationSupportDirectory inDomains:NSUserDomainMask] lastObject];
+  return [[applicationDocumentsDirectory URLByAppendingPathComponent:EXUpdatesErrorLogFile] path];
 }
 
 @end